camera/requirements.txt
accelerate==1.0.1
asttokens==3.0.1
backcall==0.2.0
bitsandbytes==0.45.5
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
debugpy==1.8.18
decorator==5.2.1
executing==2.2.1
filelock==3.16.1
fsspec==2025.3.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
packaging==25.0
panda-python==0.8.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyrealsense2==2.55.1.6486
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
timm==0.9.10
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.1
triton==3.0.0
typing_extensions==4.13.2
urllib3==2.2.3
wcwidth==0.2.14
websockets==13.1
zipp==3.20.2
